mux的各种node拆分开，准备写regex的node了
